# Conflicts:
#	.gitignore
#	README.md
#	ocr_lines_new3.py

+++ csv_handler.py
@@ -9,7 +9,7 @@ TIME_FORMAT = "%m-%d-%Y_%H.%M.%S"
 DELIMITER = "|"
 FILE_SUFFIX = ".csv"
 
-COL_NAMES = ["SuDoc", "FilteredSuDoc", "Title", "Author", "PublicationDate", "ImagePath"]
+COL_NAMES = ["SuDoc", "Status", "FilteredSuDoc", "GovernmentNumber", "Title", "Author", "PublicationDate"]
 
 
 def generate_filename() -> str:
@@ -17,81 +17,79 @@ def generate_filename() -> str:
 
 
 class SuDocRecord:
-    def __init__(self, sudoc, filtered_sudoc, title, author, publication_date, image_path):
+    def __init__(self, sudoc: str, status: str, filtered_sudoc: str, gov_num: str, title: str,
+                 author: str, publication_date: str):
         self.sudoc = sudoc
+        self.status = status
         self.filtered_sudoc = filtered_sudoc
+        self.gov_num = gov_num
         self.title = title
         self.author = author
         self.publication_date = publication_date
-        self.image_path = image_path
 
     def get_dict(self) -> dict[str]:
         record_dict = {COL_NAMES[0]: f"{self.sudoc}",
-                       COL_NAMES[1]: f"{self.filtered_sudoc}",
-                       COL_NAMES[2]: f"{self.title}",
-                       COL_NAMES[3]: f"{self.author}",
-                       COL_NAMES[4]: f"{self.publication_date}",
-                       COL_NAMES[5]: f"{self.image_path}"}
+                       COL_NAMES[1]: f"{self.status}",
+                       COL_NAMES[2]: f"{self.filtered_sudoc}",
+                       COL_NAMES[3]: f"{self.gov_num}",
+                       COL_NAMES[4]: f"{self.title}",
+                       COL_NAMES[5]: f"{self.author}",
+                       COL_NAMES[6]: f"{self.publication_date}"}
         return record_dict
 
 
 class CSVDocument:
-    def __init__(self, data_handler: fh.FileHandler, folder_path="", file_name="", file_path=""):
+    def __init__(self, data_handler: fh.FileHandler, folder_path="", file_name="", file_path="", read_only=True):
         self.__datahandler = data_handler
-        self.__file_contents: [dict] = []
-        self.__file_rows: list[SuDocRecord] = []
+        self.__file_contents: list[dict] = []
+        self.__file_rows: list[list[str]] = []
         self.__file_name = file_name
         self.__file_path = file_path
         self.__folder_path = folder_path
-        if self.__folder_path == "":
-            self.__folder_path = os.path.join(self.__datahandler.get_program_path(), FOLDER_NAME)
-            print(self.__folder_path)
+        self.__read_only = read_only
 
-        # Try Load filename
-        if self.__file_path == "" and self.__file_name != "":
-            # Check that file exists
-            self.__file_path = os.path.join(self.__folder_path, self.__file_name)
-            if not os.path.exists(self.__file_path):
-                print("File or folder does not exist... generating a new file instead")
-                self.__file_name = ""
-
-        # Generate filename
-        if self.__file_path == "" and self.__file_name == "":
-            self.__file_name = generate_filename()
-            self.__file_path = os.path.join(self.__folder_path, self.__file_name)
-            open(self.__file_path, 'w')
-
-        # Load file contents into [dict]
+        if file_path == "":
+            if self.__folder_path == "":
+                self.__folder_path = os.path.join(self.__datahandler.get_program_path(), FOLDER_NAME)
+                print(self.__folder_path)
+            # Try Load filename
+            if self.__file_name != "":
+                # Check that file exists
+                self.__file_path = os.path.join(self.__folder_path, self.__file_name)
+                if not os.path.exists(self.__file_path):
+                    print("File or folder does not exist... generating a new file instead")
+                    self.__file_name = ""
+            else:
+                self.__file_name = generate_filename()
+                self.__file_path = os.path.join(self.__folder_path, self.__file_name)
+                open(self.__file_path, 'w')
 
         # Check if file has column names
         col_names = []
         with open(self.__file_path, mode='r') as csv_file:
             csv_reader = csv.reader(csv_file, delimiter='|')
             for row in csv_reader:
-                col_names = row
-                break
+                self.__file_rows.append(row)
 
         # Check if file is not
-        if len(col_names) != 0 and col_names != COL_NAMES:
-            self.__file_path = os.path.join(self.__folder_path, generate_filename())
-            open(self.__file_path, 'w')
-            print("ERROR: FILE CONTAINS ALTERNATE DATA... closing document to avoid overwriting data.")
-            print(f"{col_names}")
+        # if len(col_names) != 0 and col_names != COL_NAMES:
+        #     self.__file_path = os.path.join(self.__folder_path, generate_filename())
+        #     open(self.__file_path, 'w')
+        #     print("ERROR: FILE CONTAINS ALTERNATE DATA... closing document to avoid overwriting data.")
+        #     print(f"{col_names}")
 
-        if len(col_names) <= 1:
-            # Insert column names
-            with open(self.__file_path, mode='w', newline='') as file:
-                csv_writer = csv.DictWriter(file, delimiter='|', fieldnames=COL_NAMES)
-                csv_writer.writeheader()
-            return
-
-        # Load dictionary with csv data
-        with open(self.__file_path, mode='r') as file:
-            csv_file = csv.DictReader(file, delimiter=DELIMITER)
-            for row in csv_file:
-                self.__file_contents.append(row)
-
-        print(self.__file_contents)
+        # if len(col_names) <= 1:
+        #     # Insert column names
+        #     with open(self.__file_path, mode='w', newline='') as file:
+        #         csv_writer = csv.DictWriter(file, delimiter='|', fieldnames=COL_NAMES)
+        #         csv_writer.writeheader()
+        #     return
+        #
+        # # Load dictionary with csv data
+        # with open(self.__file_path, mode='r') as file:
+        #     csv_file = csv.DictReader(file, delimiter=DELIMITER)
+        #     for row in csv_file:
+        #         self.__file_contents.append(row)
 
         # TODO: optional static path -> i.e. configing a folder to be the default path for csv files
         # self.__file_path = ""
@@ -124,25 +122,57 @@ class CSVDocument:
         return
 
     def add_row(self, row: dict):
+        if self.__read_only:
+            print("cant add to a read only file")
+            return
+
         self.__file_contents.append(row)
         self.__write_contents()
 
     def get_all_sudocs(self) -> list[str]:
         sudocs = []
 
-        for row in self.__file_contents:
-            sudocs.append(row["SuDoc"])
+        if len(self.__file_rows) <= 1:
+            return []
+
+        sudoc_index = -1
+        for i in range(len(self.__file_rows[0])):
+            if self.__file_rows[0][i] == "SuDoc":
+                sudoc_index = i
+                break
+
+        if sudoc_index == -1:
+            return []
+
+        for row in self.__file_rows:
+            sudocs.append(row[sudoc_index])
 
         return sudocs
 
     def __write_contents(self):
+        if self.__read_only:
+            print("cant write to a read only file")
+            return
+
         with open(self.__file_path, mode='w', newline='') as file:
             csv_writer = csv.DictWriter(file, delimiter=DELIMITER, fieldnames=COL_NAMES)
             csv_writer.writeheader()
             csv_writer.writerows(self.__file_contents)
 
     def write_row(self, sudoc_record: SuDocRecord):
+        if self.__read_only:
+            print("cant write a row to a read only file")
+            return
         return
+
+    def write_contents_to_file(self):
+        if self.__read_only:
+            return
+
+        with open(self.__file_path, mode='w', newline='') as file:
+            csv_writer = csv.DictWriter(file, delimiter=DELIMITER, fieldnames=COL_NAMES)
+            csv_writer.writeheader()
+            csv_writer.writerows(self.__file_contents)
 
     def get_file_path(self) -> str:
         return self.__file_path
@@ -157,11 +187,11 @@ class CSVDocument:
         return filename
 
 
-if __name__ == "__main__":
-    start_time = time.time()
-    csvfile = CSVDocument(fh.FileHandler(), file_name="extraction_02-18-2024_13.46.13.csv")
-    # csvfile.add_row({"SuDoc": "su", "FilteredSuDoc": "fs", "Title": "ti", "Author": "au", "PublicationDate": "pd", "ImagePath": ""})
-    print(csvfile.get_all_sudocs())
-    end_time = time.time()
-
-    print(f"Total time: {end_time - start_time}")
+# if __name__ == "__main__":
+#     start_time = time.time()
+#     csvfile = CSVDocument(fh.FileHandler(), file_name="extraction_02-18-2024_13.46.13.csv")
+#     # csvfile.add_row({"SuDoc": "su", "FilteredSuDoc": "fs", "Title": "ti", "Author": "au", "PublicationDate": "pd", "ImagePath": ""})
+#     print(csvfile.get_all_sudocs())
+#     end_time = time.time()
+#
+#     print(f"Total time: {end_time - start_time}")
